fix(pefcheck): PEFLoaderInfoHeader signed section fields + pre-relocation vector wording
LoaderInfoHeader is a fixed 56-byte / 14-field structure; mainSection,
initSection, termSection are SInt32, the rest UInt32. There is no SInt16
loader-header layout to override (the SInt16 sectionIndex belongs to the
PEFExportedSymbol table). Keep the raw big-endian u32 decode internally
and expose the three signed fields as i32; add regression tests for -1
(and that a negative non -1 mainSection is reported out-of-range, not a
panic).

Refine the transition-vector wording: raw zero/stub bytes at the
special-main location are pre-relocation contents; they do NOT establish
the vector is absent from the PEF representation, since CFM relocation
materializes pointer values at preparation time.

+++ pefcheck/src/pef.rs
@@ -35,13 +35,21 @@ impl Section {
     }
 }
 
+/// The PEF loader info header — a fixed 56-byte structure of 14 four-byte
+/// fields (Mac OS Runtime Architectures). `mainSection`, `initSection` and
+/// `termSection` are SInt32 (a value of -1 = no such symbol); all other
+/// count/offset fields are UInt32. There is no SInt16 loader-header layout
+/// to override — the SInt16 sectionIndex belongs to the PEF exported-symbol
+/// table entry (`Export::section_index`), not to the loader header. The
+/// parser keeps the raw big-endian u32 decode internally and exposes the
+/// three signed fields as `i32`.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct LoaderInfo {
-    pub main_section: u32,
+    pub main_section: i32,
     pub main_offset: u32,
-    pub init_section: u32,
+    pub init_section: i32,
     pub init_offset: u32,
-    pub term_section: u32,
+    pub term_section: i32,
     pub term_offset: u32,
     pub imported_library_count: u32,
     pub total_imported_symbol_count: u32,
@@ -143,11 +151,11 @@ impl Container {
         need(data, lo, ll, "loader container")?;
         need(data, lo, 0x38, "loader info header")?;
         let loader = LoaderInfo {
-            main_section: u32_at(data, lo),
+            main_section: u32_at(data, lo) as i32,
             main_offset: u32_at(data, lo + 4),
-            init_section: u32_at(data, lo + 8),
+            init_section: u32_at(data, lo + 8) as i32,
             init_offset: u32_at(data, lo + 12),
-            term_section: u32_at(data, lo + 16),
+            term_section: u32_at(data, lo + 16) as i32,
             term_offset: u32_at(data, lo + 20),
             imported_library_count: u32_at(data, lo + 24),
             total_imported_symbol_count: u32_at(data, lo + 28),
